+++ src/assets/scripts/client/commands/definitions/CommandModel.js
