+++ src/assets/scripts/client/commands/definitions/aircraftCommand/AircraftCommandModel.js
