+++ src/assets/scripts/client/commands/definitions/scopeCommand/scopeCommandMap.js
