+++ src/assets/scripts/client/commands/definitions/systemCommand/SystemCommandModel.js
