+++ src/assets/scripts/client/commands/definitions/systemCommand/systemCommandMap.js
@@ -1,6 +1,6 @@
-import { singleArgumentValidator, zeroArgumentsValidator, zeroOrOneArgumentValidator } from '../parsers/argumentValidators';
-import { convertStringToNumber } from '../../utilities/unitConverters';
-import { timewarpParser } from '../parsers/argumentParsers';
+import { singleArgumentValidator, zeroArgumentsValidator, zeroOrOneArgumentValidator } from '../../parsers/argumentValidators';
+import { convertStringToNumber } from '../../../utilities/unitConverters';
+import { timewarpParser } from '../../parsers/argumentParsers';
 import { noop } from '../utils';
 
 /**
