+++ src/assets/scripts/client/commands/definitions/utils.js
