+++ test/commands/definitions/aircraftCommand/AircraftCommandModel.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable arrow-parens, max-len, import/no-extraneous-dependencies */
 import ava from 'ava';
 
-import AircraftCommandModel from '../../../src/assets/scripts/client/commands/aircraftCommand/AircraftCommandModel';
+import AircraftCommandModel from '../../../../src/assets/scripts/client/commands/definitions/aircraftCommand/AircraftCommandModel';
 
 
 ava('does not thow when instantiated without parameters', t => {
